Fix tests for changes to edit/history pages
- Slice tests into smaller files for better visibility
- Test the edit page redirects to the review page
- Smoke test the document history page
