+++ ds_caselaw_editor_ui/templates/judgment/history.html
@@ -1,7 +1,7 @@
 {% extends "layouts/judgment_with_sidebar.html" %}
 {% load i18n %}
 {% block judgment_header %}
-  {% include "includes/judgment/metadata_panel_static.html" with judgment=judgment %}
+  {% include "includes/judgment/metadata_panel_static.html" with judgment=document %}
 {% endblock judgment_header %}
 {% block judgment_content %}
   <div>
@@ -13,10 +13,10 @@
         </tr>
       </thead>
       <tbody>
-        {% for version in judgment.versions %}
+        {% for version in document.versions %}
           <tr>
             <td>
-              <a href="{% url 'full-text-html' judgment_uri %}?version_uri={{ version.uri }}">Version {{ version.version }}</a>
+              <a href="{% url 'full-text-html' document_uri %}?version_uri={{ version.uri }}">Version {{ version.version }}</a>
             </td>
           </tr>
         {% endfor %}
