+++ judgments/tests/test_judgments.py
@@ -13,105 +13,6 @@ def assert_match(regex, string):
     assert re.search(regex, string) is not None
 
 
-class TestJudgmentEdit(TestCase):
-    @patch("judgments.views.judgment_edit.get_judgment_by_uri_or_404")
-    @patch("judgments.utils.api_client.document_exists")
-    @patch("judgments.utils.api_client.get_document_type_from_uri")
-    def test_judgment_edit_view(self, document_type, document_exists, mock_judgment):
-        document_type.return_value = Judgment
-        document_exists.return_value = None
-
-        judgment = JudgmentFactory.build(
-            uri="edtest/4321/123",
-            name="Test v Tested",
-        )
-        mock_judgment.return_value = judgment
-
-        self.client.force_login(User.objects.get_or_create(username="testuser")[0])
-
-        assert (
-            reverse("edit-judgment", kwargs={"judgment_uri": judgment.uri})
-            == "/edtest/4321/123/edit"
-        )
-
-        response = self.client.get(
-            reverse("edit-judgment", kwargs={"judgment_uri": judgment.uri})
-        )
-
-        decoded_response = response.content.decode("utf-8")
-        self.assertIn("Test v Tested", decoded_response)
-        assert response.status_code == 200
-        self.assertIn(
-            '<option value="UKSC">United Kingdom Supreme Court</option>',
-            decoded_response,
-        )
-
-    @patch("judgments.views.judgment_edit.api_client")
-    @patch("judgments.views.judgment_edit.get_judgment_by_uri_or_404")
-    def test_edit_judgment(self, mock_judgment, api_client):
-        judgment = JudgmentFactory.build(
-            uri="edittest/4321/123",
-            name="Test v Tested",
-        )
-        mock_judgment.return_value = judgment
-
-        self.client.force_login(User.objects.get_or_create(username="testuser")[0])
-        User.objects.get_or_create(username="testuser2")
-
-        self.client.post(
-            "/pubtest/4321/123/edit",
-            {
-                "judgment_uri": "/edittest/4321/123",
-                "metadata_name": "New Name",
-                "neutral_citation": "[4321] TEST 123",
-                "court": "Court of Testing",
-                "judgment_date": "2023-01-02",
-                "assigned_to": "testuser2",
-            },
-        )
-
-        api_client.set_judgment_name.assert_called_with(
-            "/edittest/4321/123", "New Name"
-        )
-        api_client.set_judgment_citation.assert_called_with(
-            "/edittest/4321/123", "[4321] TEST 123"
-        )
-        api_client.set_judgment_court.assert_called_with(
-            "/edittest/4321/123", "Court of Testing"
-        )
-        api_client.set_judgment_date.assert_called_with(
-            "/edittest/4321/123", "2023-01-02"
-        )
-        api_client.set_property.assert_called_with(
-            "/edittest/4321/123", "assigned-to", "testuser2"
-        )
-
-    @patch("judgments.views.judgment_edit.api_client")
-    @patch("judgments.views.judgment_edit.get_judgment_by_uri_or_404")
-    def test_edit_judgment_without_assignment(self, mock_judgment, api_client):
-        judgment = JudgmentFactory.build(
-            uri="edittest/4321/123",
-            name="Test v Tested",
-        )
-        mock_judgment.return_value = judgment
-
-        self.client.force_login(User.objects.get_or_create(username="testuser")[0])
-        User.objects.get_or_create(username="testuser2")
-
-        self.client.post(
-            "/pubtest/4321/123/edit",
-            {
-                "judgment_uri": "/edittest/4321/123",
-                "metadata_name": "New Name",
-                "neutral_citation": "[4321] TEST 123",
-                "court": "Court of Testing",
-                "judgment_date": "2023-01-02",
-            },
-        )
-
-        api_client.set_property.assert_not_called()
-
-
 class TestJudgmentView(TestCase):
     @patch("judgments.views.full_text.get_judgment_by_uri_or_404")
     @patch("judgments.utils.api_client.document_exists")
